move color picker back to qt library
