+++ qt/src/napqt/colorpicker.cpp
