+++ qt/src/napqt/colorpicker.h
